Fix BC5 compression
As I understand it, the current way of populating the decompressed
linear data is incorrect. Specifically the lack of use of dy in
the computation of the offset into the linear image.

Unless I've been using the code wrong I think there's a bug, which
this commit fixes (not sure you want to have this coding style -
feel free to change).